+++ srcs/vectors/vect_operations.c
@@ -22,7 +22,7 @@ t_vect	substract_v(t_vect a, t_vect b)
 	return (ret);
 }
 
-t_vect	mult_v(t_vect a, float t)
+t_vect	mult_v(t_vect a, double t)
 {
 	t_vect	ret;
 
@@ -42,7 +42,7 @@ t_vect	add_v(t_vect a, t_vect b)
 	return (ret);
 }
 
-float	norm_v(t_vect a)
+double	norm_v(t_vect a)
 {
 	return (sqrtf(a.x * a.x + a.y * a.y + a.z * a.z));
 }
